requirements.txt
requests>=2.31.0
pandas>=2.0.0
lxml>=4.9.0
PyYAML>=6.0
pytest>=7.0.0
urllib3>=2.0.0
python-dateutil>=2.8.2
openpyxl>=3.1.0
